# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Deetoo API
  version: 1.0.0
  description: 'Deetoo multi-vendor food delivery platform API (Kenya/East Africa).


    **Coverage note (honest, not aspirational):** this file was generated from the actual Express router
    source (scripts/openapi-coverage-check.ts enforces this in CI going forward -- it fails the build
    if a route exists in code but not here, or vice versa), so every path, method, tag, and auth requirement
    below is real. Of the 261 operations, 31 core commerce-flow operations (public discovery, auth, cart,
    checkout, order, payment) have hand-verified request/response schemas transcribed from the actual
    Zod validators and domain types. The remaining operations are marked `STUB` in their summary: they
    are registered with the correct path/method/tag/security so drift is still caught, but their request/response
    bodies are intentionally left as a generic envelope rather than guessed. Replacing each STUB with
    its real schema is tracked as ongoing work, not represented here as already done.'
  contact:
    name: Deetoo Engineering
servers:
- url: http://localhost:3000/api/v1
  description: Local development
...
security:
- bearerAuth: []
paths:
...
  /rider/earnings:
    get:
      summary: Return the authenticated Rider's earning ledger view
      tags: [Rider]
      operationId: rider_get_earnings
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Rider earnings and eligible balance }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /finance/accounts:
    get:
      summary: STUB - GET /finance/accounts (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_finance_accounts
      security:
      - bearerAuth: []
  /finance/accounts/{id}/entries:
    get:
      summary: STUB - GET /finance/accounts/:id/entries (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      operationId: finance_get_finance_accounts_id_entries
      security:
      - bearerAuth: []
  /finance/transactions:
    get:
      summary: STUB - GET /finance/transactions (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_finance_transactions
      security:
      - bearerAuth: []
  /finance/adjustments:
    get:
      summary: List financial adjustments awaiting or completing review
      tags: [Finance]
      responses:
        '200':
          description: Financial adjustments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_finance_adjustments
      security:
      - bearerAuth: []
    post:
      summary: STUB - POST /finance/adjustments (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_post_finance_adjustments
      security:
      - bearerAuth: []
  /finance/settlements:
    get:
      summary: STUB - GET /finance/settlements (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_finance_settlements
      security:
      - bearerAuth: []
  /finance/settlements/calculate:
    post:
      summary: STUB - POST /finance/settlements/calculate (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_post_finance_settlements_calculate
      security:
      - bearerAuth: []
  /finance/settlements/{id}/approve:
    post:
      summary: STUB - POST /finance/settlements/:id/approve (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      operationId: finance_post_finance_settlements_id_approve
      security:
      - bearerAuth: []
  /finance/settlements/{id}/pay:
    post:
      summary: STUB - POST /finance/settlements/:id/pay (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      operationId: finance_post_finance_settlements_id_pay
      security:
      - bearerAuth: []
  /finance/payouts:
    get:
      summary: STUB - GET /finance/payouts (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_finance_payouts
      security:
      - bearerAuth: []
  /finance/payouts/calculate:
    post:
      summary: STUB - POST /finance/payouts/calculate (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_post_finance_payouts_calculate
      security:
      - bearerAuth: []
  /finance/payouts/{id}/approve:
    post:
      summary: STUB - POST /finance/payouts/:id/approve (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      operationId: finance_post_finance_payouts_id_approve
      security:
      - bearerAuth: []
  /finance/payouts/{id}/pay:
    post:
      summary: STUB - POST /finance/payouts/:id/pay (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      operationId: finance_post_finance_payouts_id_pay
      security:
      - bearerAuth: []
  /finance/rider/earnings:
    get:
      summary: STUB - GET /finance/rider/earnings (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_finance_rider_earnings
      security:
      - bearerAuth: []
  /finance/merchant/statement:
    get:
      summary: STUB - GET /finance/merchant/statement (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_finance_merchant_statement
      security:
      - bearerAuth: []
  /finance/profitability:
    get:
      summary: STUB - GET /finance/profitability (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_finance_profitability
      security:
      - bearerAuth: []
  /finance/orders/{orderId}:
    get:
      summary: STUB - GET /finance/orders/:orderId (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: string
      operationId: finance_get_finance_orders_orderId
      security:
      - bearerAuth: []
  /finance/reconciliation/report:
    get:
      summary: STUB - GET /finance/reconciliation/report (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_get_finance_reconciliation_report
      security:
      - bearerAuth: []
  /finance/reconciliation/heal:
    post:
      summary: STUB - POST /finance/reconciliation/heal (registered, not yet individually specified)
      tags:
      - Finance
      responses:
        '200':
          description: Success (unspecified response shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_post_finance_reconciliation_heal
      security:
      - bearerAuth: []
...
  /orders/{id}:
    get:
      summary: Get an order with items, snapshots, and timeline
      tags:
      - Order
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Resource ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: order_get_orders_id
      security:
      - bearerAuth: []
...
  /finance/adjustments/{id}/approve:
    post:
      summary: Approve and post a requested financial adjustment
      tags: [Finance]
      parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
      responses:
        '200':
          description: Posted adjustment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_post_adjustments_id_approve
      security:
      - bearerAuth: []
  /finance/adjustments/{id}/reject:
    post:
      summary: Reject a requested financial adjustment
      tags: [Finance]
      parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
      responses:
        '200':
          description: Rejected adjustment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGeneric'
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: finance_post_adjustments_id_reject
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session access token issued by POST /auth/login or /auth/register/customer.
    webhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret
      description: Shared secret injected by the trusted ingress gateway after it independently validates
        the call genuinely originated from the provider (M-PESA Daraja). This is gateway authentication,
        not a provider-issued signature -- Safaricom does not sign these requests itself. Stripe webhooks
        instead use Stripe-Signature HMAC verification against the raw body and are not covered by this
        scheme.
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing, invalid, or expired credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    ApiError:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorDetail'
...
    ApiResponseGeneric:
      type: object
      required:
      - data
      description: Generic success envelope. data is intentionally untyped here where a route has not
        yet had its specific response schema individually verified (see the STUB markers in the paths
        above).
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        requestId:
          type: string
        success:
          type: boolean
...
    OrderResponse:
      allOf:
      - $ref: '#/components/schemas/ApiResponseGeneric'
      properties:
        data:
          $ref: '#/components/schemas/Order'
...